Made Collector.cpp Buildable
added pitch_Encoder_Channels A and B, fixed getPitchEncoder and
changePitch

Signed Off: Ross Bajocich

+++ Mercury2013/Subsystems/Collector.h
@@ -2,15 +2,14 @@
 #define _COLLECTOR_H
 
 #include "WPILib.h"
-#include "../Robotmap.h"
 
 class Collector : public Subsystem {
 	Encoder *collectorEncoder;
 	Encoder *pitchEncoder;
 	SpeedController *collectorMotor;
 	SpeedController *collectorPitchMotor;
-	float currentAngle;
-	float startAngle;
+	double currentAngle;
+	double startAngle;
 public:
 	Collector();
 	~Collector();
